# maanslogen-admin/openapi.json
{"openapi":"3.0.0","paths":{"/api/admin/users":{"get":{"description":"Retrieve a list of all users","operationId":"UserController_getAll","parameters":[],"responses":{"200":{"description":"List of users retrieved successfully"}},"summary":"Get all users","tags":["Admin – Users"]},"post":{"description":"Create a new user account","operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"description":"User data to create","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created successfully"},"400":{"description":"Invalid input data"}},"summary":"Create a new user","tags":["Admin – Users"]}},"/api/admin/upload/presign/{context}":{"post":{"description":"\n**Flow:**\n1. Kald dette endpoint med `context` (beverage-images, user-profile eller category-icons). Bucket og mappesti bestemmes i backend.\n2. For hver slot: upload fil med **PUT** til `uploadUrl`, body = fil, header `Content-Type: image/jpeg` (eller png/gif/webp).\n3. Brug den returnerede `url` for hver slot når du opretter entiteten (Beverage/User/Category) med `images: [ { url, type } ]`.\n\nPresigned URLs udløber efter 15 minutter (kan overstyres med `expiresInSeconds`).\n    ","operationId":"UploadController_presign","parameters":[{"name":"context","required":true,"in":"path","description":"Upload-type – backend vælger bucket og mappesti","schema":{"enum":["beverage-images","user-profile","category-icons"],"type":"string"}},{"name":"x-test-bucket","required":true,"in":"header","schema":{"type":"string"}},{"name":"X-Test-Bucket","in":"header","description":"Valgfri bucket (fx til test) – opretter ny bucket med public policy hvis den ikke findes","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresignUploadDto"}}}},"responses":{"201":{"description":"Presigned URLs oprettet","content":{"application/json":{"schema":{"type":"object","properties":{"uploads":{"type":"array","items":{"type":"object","properties":{"uploadUrl":{"type":"string","description":"PUT fil til denne URL"},"url":{"type":"string","description":"Brug denne URL i images[] ved oprettelse"},"key":{"type":"string"},"type":{"type":"string","enum":["THUMBNAIL","LARGE","PROFILE","ICON"]},"width":{"type":"number","description":"Bredde (px) – brug ved oprettelse"},"height":{"type":"number","description":"Højde (px) – brug ved oprettelse"},"bucket":{"type":"string"}}}}}}}}},"400":{"description":"Ugyldig request eller bucket findes ikke"}},"summary":"Get presigned upload URLs (bucket og sti sættes i backend)","tags":["Admin – Upload"]}},"/api/admin/beverages":{"get":{"operationId":"BeverageAdminController_getAll","parameters":[],"responses":{"200":{"description":"List of beverages"}},"summary":"[Admin] Get all beverages","tags":["Admin – Beverages"]},"post":{"operationId":"BeverageAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBeverageDto"}}}},"responses":{"201":{"description":"Beverage created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a new beverage","tags":["Admin – Beverages"]}},"/api/admin/beverages/{id}":{"get":{"operationId":"BeverageAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Beverage ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Beverage found"},"404":{"description":"Beverage not found"}},"summary":"[Admin] Get beverage by ID","tags":["Admin – Beverages"]}},"/api/beverages":{"get":{"operationId":"BeverageWebController_getAll","parameters":[],"responses":{"200":{"description":"List of beverages"}},"summary":"[Web] Get all beverages (public)","tags":["Web – Beverages"]}},"/api/beverages/{id}":{"get":{"operationId":"BeverageWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Beverage ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Beverage found"},"404":{"description":"Beverage not found"}},"summary":"[Web] Get beverage by ID (public)","tags":["Web – Beverages"]}},"/api/admin/reviews":{"get":{"operationId":"ReviewAdminController_getAll","parameters":[],"responses":{"200":{"description":"List of reviews"}},"summary":"[Admin] Get all reviews","tags":["Admin – Reviews"]},"post":{"operationId":"ReviewAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":"Review created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a review","tags":["Admin – Reviews"]}},"/api/admin/reviews/{id}":{"get":{"operationId":"ReviewAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Review ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Review found"},"404":{"description":"Not found"}},"summary":"[Admin] Get review by ID","tags":["Admin – Reviews"]}},"/api/reviews":{"get":{"operationId":"ReviewWebController_getAll","parameters":[],"responses":{"200":{"description":"List of reviews"}},"summary":"[Web] Get all reviews (public)","tags":["Web – Reviews"]}},"/api/reviews/{id}":{"get":{"operationId":"ReviewWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Review ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Review found"},"404":{"description":"Not found"}},"summary":"[Web] Get review by ID (public)","tags":["Web – Reviews"]}},"/api/admin/beverage-categories":{"get":{"operationId":"BeverageCategoryAdminController_getAll","parameters":[],"responses":{"200":{"description":"List of categories"}},"summary":"[Admin] Get all beverage categories","tags":["Admin – Beverage Categories"]},"post":{"operationId":"BeverageCategoryAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"Category created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a beverage category","tags":["Admin – Beverage Categories"]}},"/api/admin/beverage-categories/{id}":{"get":{"operationId":"BeverageCategoryAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Category found"},"404":{"description":"Not found"}},"summary":"[Admin] Get category by ID","tags":["Admin – Beverage Categories"]}},"/api/beverage-categories":{"get":{"operationId":"BeverageCategoryWebController_getAll","parameters":[],"responses":{"200":{"description":"List of categories"}},"summary":"[Web] Get all beverage categories (public)","tags":["Web – Beverage Categories"]}},"/api/beverage-categories/{id}":{"get":{"operationId":"BeverageCategoryWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Category found"},"404":{"description":"Not found"}},"summary":"[Web] Get category by ID (public)","tags":["Web – Beverage Categories"]}},"/api/admin/beverage-types":{"get":{"operationId":"BeverageTypeAdminController_getAll","parameters":[],"responses":{"200":{"description":"List of types"}},"summary":"[Admin] Get all beverage types","tags":["Admin – Beverage Types"]},"post":{"operationId":"BeverageTypeAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBeverageTypeDto"}}}},"responses":{"201":{"description":"Type created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a beverage type","tags":["Admin – Beverage Types"]}},"/api/admin/beverage-types/{id}":{"get":{"operationId":"BeverageTypeAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Beverage type ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Type found"},"404":{"description":"Not found"}},"summary":"[Admin] Get beverage type by ID","tags":["Admin – Beverage Types"]}},"/api/beverage-types":{"get":{"operationId":"BeverageTypeWebController_getAll","parameters":[],"responses":{"200":{"description":"List of types"}},"summary":"[Web] Get all beverage types (public)","tags":["Web – Beverage Types"]}},"/api/beverage-types/{id}":{"get":{"operationId":"BeverageTypeWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Beverage type ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Type found"},"404":{"description":"Not found"}},"summary":"[Web] Get beverage type by ID (public)","tags":["Web – Beverage Types"]}},"/api/admin/brands":{"get":{"operationId":"BrandAdminController_getAll","parameters":[],"responses":{"200":{"description":"List of brands"}},"summary":"[Admin] Get all brands","tags":["Admin – Brands"]},"post":{"operationId":"BrandAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBrandDto"}}}},"responses":{"201":{"description":"Brand created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a brand","tags":["Admin – Brands"]}},"/api/admin/brands/{id}":{"get":{"operationId":"BrandAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Brand ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Brand found"},"404":{"description":"Not found"}},"summary":"[Admin] Get brand by ID","tags":["Admin – Brands"]}},"/api/brands":{"get":{"operationId":"BrandWebController_getAll","parameters":[],"responses":{"200":{"description":"List of brands"}},"summary":"[Web] Get all brands (public)","tags":["Web – Brands"]}},"/api/brands/{id}":{"get":{"operationId":"BrandWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Brand ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Brand found"},"404":{"description":"Not found"}},"summary":"[Web] Get brand by ID (public)","tags":["Web – Brands"]}},"/api/admin/attributes":{"post":{"operationId":"AttributeDefinitionAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAttributeDefinitionDto"}}}},"responses":{"201":{"description":"Attribute definition created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create attribute definition","tags":["Admin – Attribute Definitions"]},"get":{"operationId":"AttributeDefinitionAdminController_findAll","parameters":[],"responses":{"200":{"description":"List of attribute definitions"}},"summary":"[Admin] Get all attribute definitions","tags":["Admin – Attribute Definitions"]}},"/api/admin/attributes/category/{categoryId}":{"get":{"operationId":"AttributeDefinitionAdminController_findByCategory","parameters":[{"name":"categoryId","required":true,"in":"path","description":"Beverage category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of attribute definitions"}},"summary":"[Admin] Get attribute definitions by category","tags":["Admin – Attribute Definitions"]}},"/api/admin/attributes/{id}":{"get":{"operationId":"AttributeDefinitionAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Attribute definition ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Attribute definition found"},"404":{"description":"Not found"}},"summary":"[Admin] Get attribute definition by ID","tags":["Admin – Attribute Definitions"]}},"/api/attributes":{"get":{"operationId":"AttributeDefinitionWebController_findAll","parameters":[],"responses":{"200":{"description":"List of attribute definitions"}},"summary":"[Web] Get all attribute definitions (public)","tags":["Web – Attribute Definitions"]}},"/api/attributes/category/{categoryId}":{"get":{"operationId":"AttributeDefinitionWebController_findByCategory","parameters":[{"name":"categoryId","required":true,"in":"path","description":"Beverage category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of attribute definitions"}},"summary":"[Web] Get attribute definitions by category (public)","tags":["Web – Attribute Definitions"]}},"/api/attributes/{id}":{"get":{"operationId":"AttributeDefinitionWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Attribute definition ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Attribute definition found"},"404":{"description":"Not found"}},"summary":"[Web] Get attribute definition by ID (public)","tags":["Web – Attribute Definitions"]}},"/api/admin/questions":{"post":{"operationId":"QuestionAdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionDto"}}}},"responses":{"201":{"description":"Question created"},"400":{"description":"Invalid input"}},"summary":"[Admin] Create a question","tags":["Admin – Questions"]},"get":{"operationId":"QuestionAdminController_findAll","parameters":[],"responses":{"200":{"description":"List of questions"}},"summary":"[Admin] Get all questions","tags":["Admin – Questions"]}},"/api/admin/questions/category/{categoryId}":{"get":{"operationId":"QuestionAdminController_findByCategory","parameters":[{"name":"categoryId","required":true,"in":"path","description":"Beverage category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of questions"}},"summary":"[Admin] Get questions by category","tags":["Admin – Questions"]}},"/api/admin/questions/type/{typeId}":{"get":{"operationId":"QuestionAdminController_findByType","parameters":[{"name":"typeId","required":true,"in":"path","description":"Beverage type ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of questions"}},"summary":"[Admin] Get questions by type","tags":["Admin – Questions"]}},"/api/admin/questions/{id}":{"get":{"operationId":"QuestionAdminController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Question ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Question found"},"404":{"description":"Not found"}},"summary":"[Admin] Get question by ID","tags":["Admin – Questions"]},"delete":{"operationId":"QuestionAdminController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Question ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Question deleted"},"404":{"description":"Not found"}},"summary":"[Admin] Delete question and compact sortOrder","tags":["Admin – Questions"]}},"/api/questions":{"get":{"operationId":"QuestionWebController_findAll","parameters":[],"responses":{"200":{"description":"List of questions"}},"summary":"[Web] Get all questions (public)","tags":["Web – Questions"]}},"/api/questions/category/{categoryId}":{"get":{"operationId":"QuestionWebController_findByCategory","parameters":[{"name":"categoryId","required":true,"in":"path","description":"Beverage category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of questions"}},"summary":"[Web] Get questions by category (public)","tags":["Web – Questions"]}},"/api/questions/type/{typeId}":{"get":{"operationId":"QuestionWebController_findByType","parameters":[{"name":"typeId","required":true,"in":"path","description":"Beverage type ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of questions"}},"summary":"[Web] Get questions by type (public)","tags":["Web – Questions"]}},"/api/questions/{id}":{"get":{"operationId":"QuestionWebController_getById","parameters":[{"name":"id","required":true,"in":"path","description":"Question ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Question found"},"404":{"description":"Not found"}},"summary":"[Web] Get question by ID (public)","tags":["Web – Questions"]}}},"info":{"title":"Maanslogen API","description":"API til Maanslogen admin og frontend","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateImageDto":{"type":"object","properties":{"url":{"type":"string","description":"URL to the image","example":"https://example.com/image.jpg"},"type":{"type":"string","description":"Type of image (e.g. thumbnail, large, profile, icon)","enum":["THUMBNAIL","LARGE","PROFILE","ICON"],"example":"THUMBNAIL"},"width":{"type":"number","description":"Bredde (px)","minimum":1,"maximum":4096},"height":{"type":"number","description":"Højde (px)","minimum":1,"maximum":4096}},"required":["url","type"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Unique username for the user","example":"alice"},"email":{"type":"string","description":"Email address of the user","example":"alice@example.com"},"passwordHash":{"type":"string","description":"Hashed password for the user","example":"hashedpassword123"},"images":{"description":"List of images (e.g. profile picture)","type":"array","items":{"$ref":"#/components/schemas/CreateImageDto"}}},"required":["username","email","passwordHash"]},"PresignSlotDto":{"type":"object","properties":{"type":{"type":"string","description":"Image type (e.g. thumbnail, large, profile, icon)","enum":["THUMBNAIL","LARGE","PROFILE","ICON"],"example":"THUMBNAIL"},"width":{"type":"number","description":"Bredde (px) – bruges i key som .../id/widthxheight. Default per type hvis udeladt.","minimum":1,"maximum":4096},"height":{"type":"number","description":"Højde (px) – bruges i key som .../id/widthxheight. Default per type hvis udeladt.","minimum":1,"maximum":4096}},"required":["type"]},"PresignUploadDto":{"type":"object","properties":{"uploads":{"description":"List of image slots – one entry per file you will upload","example":[{"type":"THUMBNAIL"},{"type":"LARGE"}],"type":"array","items":{"$ref":"#/components/schemas/PresignSlotDto"}},"expiresInSeconds":{"type":"number","description":"Presigned URL expiry in seconds (default: 900)","minimum":60,"maximum":86400}},"required":["uploads"]},"CreateBeverageDto":{"type":"object","properties":{"beverageTypeId":{"type":"string","description":"ID of the beverage type","example":"123e4567-e89b-12d3-a456-426614174000"},"brandId":{"type":"string","description":"ID of the brand","example":"123e4567-e89b-12d3-a456-426614174000"},"name":{"type":"string","description":"Name of the beverage","example":"Classic Lager"},"country":{"type":"string","description":"Country code where the beverage is from","example":"DK"},"metadata":{"type":"object","description":"Additional metadata as JSON object","example":{"notes":"Light and refreshing"}},"images":{"description":"List of images (e.g. thumbnail, large)","type":"array","items":{"$ref":"#/components/schemas/CreateImageDto"}}},"required":["beverageTypeId","brandId","name"]},"CreateReviewDto":{"type":"object","properties":{"userId":{"type":"string","description":"ID of the user creating the review","example":"123e4567-e89b-12d3-a456-426614174000"},"beverageId":{"type":"string","description":"ID of the beverage being reviewed","example":"123e4567-e89b-12d3-a456-426614174000"},"rating":{"type":"number","description":"Rating from 0 to 5","example":4.5,"minimum":0,"maximum":5},"title":{"type":"string","description":"Title of the review","example":"Great beer!"},"description":{"type":"string","description":"Detailed review description","example":"This beer has a great taste and aroma."}},"required":["userId","beverageId","rating"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the beverage category","example":"Beer"},"description":{"type":"string","description":"The description of the beverage category","example":"All types of beer"},"images":{"description":"List of images (e.g. icon)","type":"array","items":{"$ref":"#/components/schemas/CreateImageDto"}}},"required":["name","description"]},"CreateBeverageTypeDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the beverage type","example":"IPA"},"categoryId":{"type":"string","description":"ID of the beverage category this type belongs to","example":"123e4567-e89b-12d3-a456-426614174000"},"description":{"type":"string","description":"Description of the beverage type","example":"Hoppy and bitter"},"active":{"type":"boolean","description":"Whether the beverage type is active","example":true,"default":true}},"required":["name","categoryId"]},"CreateBrandDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the brand","example":"Carlsberg"},"description":{"type":"string","description":"Description of the brand","example":"Danish brewery"},"active":{"type":"boolean","description":"Whether the brand is active","example":true,"default":true}},"required":["name"]},"CreateAttributeDefinitionDto":{"type":"object","properties":{"categoryId":{"type":"string","description":"ID of the beverage category","example":"123e4567-e89b-12d3-a456-426614174000"},"typeId":{"type":"string","description":"ID of the beverage type (optional, null for category-wide attributes)","example":"123e4567-e89b-12d3-a456-426614174000"},"attributeKey":{"type":"string","description":"Unique key identifier for the attribute","example":"abv"},"displayName":{"type":"string","description":"Display name for the attribute","example":"Alkohol %"},"dataType":{"type":"string","description":"Data type of the attribute value","example":"number","enum":["string","number","boolean"]},"filterable":{"type":"boolean","description":"Whether this attribute can be used for filtering","example":true,"default":false},"required":{"type":"boolean","description":"Whether this attribute is required","example":false,"default":false}},"required":["categoryId","attributeKey","displayName","dataType"]},"CreateQuestionDto":{"type":"object","properties":{"categoryId":{"type":"string","description":"ID of the beverage category","example":"123e4567-e89b-12d3-a456-426614174000"},"typeId":{"type":"string","description":"ID of the beverage type (optional – question applies to all types in category if omitted)","example":"123e4567-e89b-12d3-a456-426614174000"},"questionText":{"type":"string","description":"Question text","example":"Hvordan er aromaen?"},"answerType":{"type":"string","description":"Answer type","example":"text","enum":["text","number","select","rating"]},"options":{"type":"object","description":"Options for select-type (e.g. array of choices)","example":["Lys","Mørk"]},"sortOrder":{"type":"number","description":"Sort order (lower = first)","minimum":0},"required":{"type":"boolean","description":"Whether the question is required","default":false}},"required":["categoryId","questionText","answerType"]}}}}