# maanslogen-api/src/modules/beverage/beverage-web.controller.ts
import { Controller, Get, Param } from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiParam } from '@nestjs/swagger';
import { BeverageService } from './beverage.service';
import { Beverage } from './dto/beverage-response.dto';

@ApiTags('Web – Beverages')
@Controller('beverages')
export class BeverageWebController {
  constructor(private beverageService: BeverageService) {}

  @Get()
  @ApiOperation({ summary: '[Web] Get all beverages (public)' })
  @ApiResponse({ status: 200, description: 'List of beverages', type: [Beverage] })
  getAll() {
    return this.beverageService.getAll();
  }

  @Get(':id')
  @ApiOperation({ summary: '[Web] Get beverage by ID (public)' })
  @ApiParam({ name: 'id', description: 'Beverage ID' })
  @ApiResponse({ status: 200, description: 'Beverage found', type: Beverage })
  @ApiResponse({ status: 404, description: 'Beverage not found' })
  getById(@Param('id') id: string) {
    return this.beverageService.getById(id);
  }
}